fix move styles files
