+++ src/core/createStyle.ts
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 import { mainStyles } from './mainStyles';
-import { Styles, StyleKeys } from './types';
+import { Styles, StyleKeys } from '../types';
 
 export const s = (...classes: StyleKeys[]): Styles => {
   const flattenClasses = classes.reduce<Styles[]>((pv, cv) => [...pv, mainStyles[cv]], []);
