--- src/core/types.ts
@@ -1,5 +1,0 @@
-export type Valueof<T> = T[keyof T]
-
-export type StyleKeys = keyof {}
-export type StyleValues = Valueof<{}>
-export type Styles = Record<StyleKeys, StyleValues>
